#2147 made kiwixMainActivity use fragment
